requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
marshmallow==3.5.0
psycopg2-binary==2.8.4